Simplify the schizo framework
Only allow one personality per submitted job, picking the one
that either correlates to the proxy cmd being used, or where
the PRRTE tool is in the install prefix location of that proxy,
or based on component priority where no other options are available.

Make every component provide a full set of options so they can fully
configure the cmd line.
