--- src/mca/schizo/pmix/Makefile.am
@@ -1,37 +1,0 @@
-#
-# Copyright (c) 2015-2020 Intel, Inc.  All rights reserved.
-# Copyright (c) 2017      IBM Corporation.  All rights reserved.
-# Copyright (c) 2020      Cisco Systems, Inc.  All rights reserved
-# $COPYRIGHT$
-#
-# Additional copyrights may follow
-#
-# $HEADER$
-#
-
-sources = \
-          schizo_pmix_component.c \
-          schizo_pmix.h \
-          schizo_pmix.c
-
-# Make the output library in this directory, and name it either
-# mca_<type>_<name>.la (for DSO builds) or libmca_<type>_<name>.la
-# (for static builds).
-
-if MCA_BUILD_prte_schizo_pmix_DSO
-component_noinst =
-component_install = mca_schizo_pmix.la
-else
-component_noinst = libmca_schizo_pmix.la
-component_install =
-endif
-
-mcacomponentdir = $(prtelibdir)
-mcacomponent_LTLIBRARIES = $(component_install)
-mca_schizo_pmix_la_SOURCES = $(sources)
-mca_schizo_pmix_la_LDFLAGS = -module -avoid-version
-mca_schizo_pmix_la_LIBADD = $(top_builddir)/src/libprrte.la
-
-noinst_LTLIBRARIES = $(component_noinst)
-libmca_schizo_pmix_la_SOURCES = $(sources)
-libmca_schizo_pmix_la_LDFLAGS = -module -avoid-version
