--- src/mca/schizo/slurm/owner.txt
@@ -1,7 +1,0 @@
-#
-# owner/status file
-# owner: institution that is responsible for this package
-# status: e.g. active, maintenance, unmaintained
-#
-owner: INTEL
-status: active
