+++ src/prted/Makefile.am
@@ -13,6 +13,7 @@
 # Copyright (c) 2014-2020 Cisco Systems, Inc.  All rights reserved
 # Copyright (c) 2015-2020 Intel, Inc.  All rights reserved.
 # Copyright (c) 2018      IBM Corporation.  All rights reserved.
+# Copyright (c) 2021      Nanook Consulting.  All rights reserved.
 # $COPYRIGHT$
 #
 # Additional copyrights may follow
@@ -28,6 +29,7 @@ headers += \
 	prted/prted.h
 
 libprrte_la_SOURCES += \
-        prted/prted_comm.c
+        prted/prted_comm.c \
+        prted/prte_app_parse.c
 
 include prted/pmix/Makefile.am
