+++ src/tools/prte/main.c
@@ -1,0 +1,35 @@
+/***************************************************************************
+ *                                                                         *
+ *          PRTE: PMIx Reference RunTime Environment (PRTE)              *
+ *                                                                         *
+ *                   https://github.com/openpmix/prte                     *
+ *                                                                         *
+ ***************************************************************************/
+
+#include "prte.h"
+
+int main(int argc, char *argv[])
+{
+    return prte(argc, argv);
+}
+
+/*
+ * Copyright (c) 2004-2005 The Trustees of Indiana University and Indiana
+ *                         University Research and Technology
+ *                         Corporation.  All rights reserved.
+ * Copyright (c) 2004-2005 The University of Tennessee and The University
+ *                         of Tennessee Research Foundation.  All rights
+ *                         reserved.
+ * Copyright (c) 2004-2005 High Performance Computing Center Stuttgart,
+ *                         University of Stuttgart.  All rights reserved.
+ * Copyright (c) 2004-2005 The Regents of the University of California.
+ *                         All rights reserved.
+ * Copyright (c) 2017-2020 Intel, Inc.  All rights reserved.
+ * Copyright (c) 2020      Cisco Systems, Inc.  All rights reserved.
+ * Copyright (c) 2021      Nanook Consulting.  All rights reserved.
+ * $COPYRIGHT$
+ *
+ * Additional copyrights may follow
+ *
+ * $HEADER$
+ */
